requirements.txt
# Production dependencies
openai>=1.0.0
numpy>=1.24.0

# Optional: load .env from project root (agent/config.py will use it if installed)
python-dotenv>=1.0.0
